Debug panel, edited countermeasures music, and more
For the debug panel, added a whole bunch of getters to print out variables.
If ball is saved, doesn't switch the music upon auto-launching the next ball.
Increased launcher force.
Added a way to resume music after a mode's been completed.
Added some in-progress colliders to help the ball go up the ramps.